old_code-AIE/pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "omegaup-editorial-generator"
version = "0.1.0"
description = "AI-powered editorial generator for omegaUp"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
omegaup-editorial-generator = "editorial_generator.cli:main"
